requirements-dev.txt
-r requirements.txt
jupyter
pylint-django
pylint-plugin-utils
autopep8
pycodestyle